tree2: Export NodeFromSchema and clarify TreeListNode.inline (#18521)
## Description

Export NodeFromSchema (which is still useful for structural lists and
maps) and clarify TreeListNode.inline.
